requirements.txt
bokeh==2.0.0
dateutils==0.6.7
django-crispy-forms==1.9.0
django-extensions==2.2.8
django-solo==1.1.3
django-summernote==0.8.11.6
django-tables2==2.2.1
gunicorn==20.0.4
ipdb==0.13.2
jupyterlab==2.0.1
pandas-interactive-html==0.1.0
pipdeptree==0.13.2
psycopg2-binary==2.8.5
requests==2.23.0
scikit-learn==0.22.2.post1
sh==1.12.14
wheel==0.34.2
xlwt==1.3.0
django-heroku==0.3.1
